refac: Predicate pushdown on reading time_series_points (#1678)
* refac: Predicate pushdown on reading time_series_points

* ...

* ...

* ...

* ...

* ...

* ...

* ...

* fix failing tests

* more fix
