+++ source/databricks/calculation_engine/package/calculation/calculation.py
@@ -40,9 +40,9 @@ def execute(args: CalculatorArgs, prepared_data_reader: PreparedDataReader) -> N
     )
 
     metering_point_time_series = prepared_data_reader.get_metering_point_time_series(
-        metering_point_periods_df,
         args.batch_period_start_datetime,
         args.batch_period_end_datetime,
+        metering_point_periods_df,
     ).cache()
 
     basis_data_writer = BasisDataWriter(args.wholesale_container_path, args.batch_id)
