+++ source/databricks/calculation_engine/package/calculation_input/table_reader.py
@@ -12,6 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from datetime import datetime
+
 from pyspark.sql import DataFrame, SparkSession
 from pyspark.sql.functions import concat_ws, col, when, lit
 
@@ -48,7 +50,8 @@ class TableReader:
             self._time_series_points_table_name = time_series_points_table_name
 
     def read_metering_point_periods(self) -> DataFrame:
-        df = self._read_table("metering_point_periods")
+        path = f"{self._calculation_input_path}/metering_point_periods"
+        df = self._spark.read.format("delta").load(path)
 
         assert_schema(df.schema, metering_point_period_schema)
 
@@ -56,15 +59,25 @@ class TableReader:
         df = self._fix_metering_point_type(df)
         return df
 
-    def read_time_series_points(self) -> DataFrame:
-        df = self._read_table(self._time_series_points_table_name)
+    def read_time_series_points(
+        self, period_start_datetime: datetime, period_end_datetime: datetime
+    ) -> DataFrame:
+        df = (
+            self._spark.read.format("delta")
+            .load(
+                f"{self._calculation_input_path}/{self._time_series_points_table_name}"
+            )
+            .where(col(Colname.observation_time) >= period_start_datetime)
+            .where(col(Colname.observation_time) < period_end_datetime)
+        )
 
         assert_schema(df.schema, time_series_point_schema)
 
         return df
 
     def read_charge_links_periods(self) -> DataFrame:
-        df = self._read_table(paths.CHARGE_LINK_PERIODS_TABLE_NAME)
+        path = f"{self._calculation_input_path}/{paths.CHARGE_LINK_PERIODS_TABLE_NAME}"
+        df = self._spark.read.format("delta").load(path)
 
         assert_schema(df.schema, charge_link_periods_schema)
 
@@ -72,7 +85,8 @@ class TableReader:
         return df
 
     def read_charge_master_data_periods(self) -> DataFrame:
-        df = self._read_table(paths.CHARGE_MASTER_DATA_PERIODS_TABLE_NAME)
+        path = f"{self._calculation_input_path}/{paths.CHARGE_MASTER_DATA_PERIODS_TABLE_NAME}"
+        df = self._spark.read.format("delta").load(path)
 
         assert_schema(df.schema, charge_master_data_periods_schema)
 
@@ -80,16 +94,13 @@ class TableReader:
         return df
 
     def read_charge_price_points(self) -> DataFrame:
-        df = self._read_table(paths.CHARGE_PRICE_POINTS_TABLE_NAME)
+        path = f"{self._calculation_input_path}/{paths.CHARGE_PRICE_POINTS_TABLE_NAME}"
+        df = self._spark.read.format("delta").load(path)
 
         assert_schema(df.schema, charge_price_points_schema)
 
         df = self._add_charge_key_column(df)
         return df
-
-    def _read_table(self, folder_name: str) -> DataFrame:
-        path = f"{self._calculation_input_path}/{folder_name}"
-        return self._spark.read.format("delta").load(path)
 
     def _add_charge_key_column(self, charge_df: DataFrame) -> DataFrame:
         return charge_df.withColumn(
@@ -184,7 +195,8 @@ class TableReader:
             # The otherwise is to avoid changing the nullability of the column.
         )
 
-    def _fix_settlement_method(self, df: DataFrame) -> DataFrame:
+    @staticmethod
+    def _fix_settlement_method(df: DataFrame) -> DataFrame:
         return df.withColumn(
             Colname.settlement_method,
             when(
